Add TimeStepper API, tests
